Fixed deployment on Vercel and modularized Swagger configuration
