+++ src/main.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { ValidationPipe } from '@nestjs/common';
+import { setupSwagger } from './swagger/swagger.setup';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -18,15 +18,7 @@ async function bootstrap() {
   );
   app.enableCors()
 
-  const config = new DocumentBuilder()
-    .setTitle('Jiru API Documentation')
-    .setDescription(' The Jiru API Documentation for the Jiru App.')
-    .setVersion('1.5')
-    .addBearerAuth()
-    .build()
-  const document = SwaggerModule.createDocument(app, config);
-
-  SwaggerModule.setup('api/docs', app, document);
+  setupSwagger(app);
 
   await app.listen(parseInt(process.env.PORT) || 3000);
 }
